feat(login): add a signup form
[Delivers #162948803]
